Separated main.js to modules, updated pomodoro
