+++ scripts/helper.js
@@ -1,0 +1,15 @@
+/*----------------------------- HELPER FUNCTIONS -----------------------------*/
+export const helper = {
+    /**
+     * Adds functions for the emotional support creature.
+     * 
+     * @function summon - Summon the creature.
+     * 
+     * TODO: Have helper give reminders.
+     * TODO: Have helper give advice.
+     * TODO: Have helper tell the user to take breaks.
+     */
+    summon: function() {
+        let phoenix = document.getElementById("phoenix");
+    }
+};
